+++ src/models/rigid_wasabi_model.py
@@ -1,3 +1,4 @@
+import os
 import gc
 import logging
 import numpy as np
@@ -21,7 +22,7 @@ class rigid_wasabi_model(pipeline_pytorch_model):
         self.name = 'rigid wasabi'
         self.version = '0.1'
         self.feature_type = MapUnitType.POLYGON
-        self.checkpoint = '/projects/bbym/shared/models/rigid_wasabi/SWIN_jaccard.ckpt'
+        self.checkpoint = 'rigid_wasabi-0.1.ckpt'
         self.args = SimpleNamespace(model='swin', superpixel='', edge=False)
         self.est_patches_per_sec = 250 # seconds
         
@@ -33,8 +34,9 @@ class rigid_wasabi_model(pipeline_pytorch_model):
         self.unpatch_mode = 'discard'
         
     #@override
-    def load_model(self):
-        self.model = SegmentationModel.load_from_checkpoint(checkpoint_path=self.checkpoint, args=self.args)
+    def load_model(self, model_dir):
+        model_path = os.path.join(model_dir, self._checkpoint) 
+        self.model = SegmentationModel.load_from_checkpoint(checkpoint_path=model_path, args=self.args)
         self.model.eval()
         self.model.to(self.device)